backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.0.14
django-axes==8.3.1
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.1
django-guardian==3.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-slugify==8.0.4
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
stripe==14.3.0
text-unidecode==1.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
